implement saving a new request
