+++ client/src/components/Form/components/index.js
@@ -2,6 +2,7 @@ import Actions from './Actions';
 import Content from './Content';
 import ContentText from './ContentText';
 import Input from './Input';
+import LoadingIndicator from './LoadingIndicator';
 import Text from './Text';
 
 export {
@@ -9,5 +10,6 @@ export {
     Content,
     ContentText,
     Input,
+    LoadingIndicator,
     Text,
 };
