+++ client/src/reducers/users.js
@@ -1,0 +1,3 @@
+import { defaultUsers } from '../constant';
+
+export default (users = defaultUsers, action) => users;
